move all tests to dedicated __tests__ folders
